move wallet classes into wallet.py
